+++ javascript/game/Camera/CameraTransformHandler.js
@@ -1,26 +1,57 @@
 import * as THREE from "three";
 
 class CameraTransformHandler {
-    constructor(mainCamera, orbitControls, cameraController) {
+    constructor(props) {
         // Fields
-        this.camera = mainCamera;
+        this.camera = props.mainCamera;
+        this.cameraState = props.cameraState;
         this.lastPosition = new THREE.Vector3();
+        this.lastRelativeDistance = 5;
+        this.lastDirection = new THREE.Vector3();
 
         // Components
-        this.orbitControls = orbitControls;
-        this.cameraController = cameraController;
+        this.orbitControls = props.orbitControls;
+        this.cameraController = props.cameraController;
     }
 
     CaptureCameraLastPosition() {
-        this.lastPosition.copy(this.camera.GetPosition()).sub(this.cameraController.GetViewTargetPosition());
+        const cameraDirection = this.camera.GetPosition().clone().sub(this.cameraController.GetViewTargetPosition().clone());
+        const newCameraPosition = this.cameraController.GetViewTargetPosition().clone().add(cameraDirection);
+
+        this.lastPosition.copy(newCameraPosition);
+        this.lastDirection.copy(cameraDirection.normalize());
+    }
+
+    CaptureCameraLastRelativeDistance() {
+        const cameraDirection = this.camera.GetPosition().clone().sub(this.cameraController.GetViewTargetPosition().clone());
+        const relativeDistance = cameraDirection.length();
+        this.lastRelativeDistance = relativeDistance;
+    }
+
+    MaintainRelativeOrbitAroundTarget() {
+        const direction = new THREE.Vector3().subVectors(
+            this.camera.GetPosition(),
+            this.cameraController.GetViewTargetPosition().clone()
+        ).normalize();
+
+        const newPosition = direction
+            .multiplyScalar(this.lastRelativeDistance)
+            .add(this.cameraController.GetViewTargetPosition());
+
+        this.camera.SetPosition(newPosition);
     }
 
     FollowTarget() {
-        if (!this.cameraController.IsControllerInteracting()) {
-            this.camera.SetPosition(this.cameraController.GetViewTargetPosition()).add(this.lastPosition);
+        if (!this.cameraController.IsControllerInteracting() && !this.cameraState.isZooming) {
+            const relativePosition = this.cameraController
+                .GetViewTargetPosition().clone()
+                .add(this.lastDirection.clone().multiplyScalar(this.lastRelativeDistance));
+            this.camera.SetPosition(relativePosition);
         }
 
         this.CaptureCameraLastPosition();
+        this.MaintainRelativeOrbitAroundTarget();
+
         this.orbitControls.target.copy(this.cameraController.GetViewTargetPosition());
     }