+++ javascript/user-interface/time-control/TimeControlRevealButton.jsx
@@ -1,9 +1,9 @@
-import { TimeControlButton } from "./TimeControlButton";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faCaretUp, faCaretDown } from "@fortawesome/free-solid-svg-icons";
-import { SolarSystemVisualizer } from "../../SolarSystemVisualizer.js";
 import { EventMediator } from "../mediator/EventMediator.js";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { GameManager } from "../../game/GameManager.js";
+import { SolarSystemVisualizer } from "../../SolarSystemVisualizer.js";
+import { TimeControlButton } from "./TimeControlButton";
+import { faCaretUp, faCaretDown } from "@fortawesome/free-solid-svg-icons";
 
 class TimeControlRevealButton extends TimeControlButton {
     constructor() {
@@ -13,6 +13,16 @@ class TimeControlRevealButton extends TimeControlButton {
             isHidden: true
         };
     }
+
+    /* -------------------------------------------------------------------------- */
+    /*                              Lifecycle Methods                             */
+    /* -------------------------------------------------------------------------- */
+
+    componentDidMount() {
+        GameManager.gameObserver.Subscribe("OnMouseDown", this.OnMouseDown.bind(this));
+        this.eventMediator = SolarSystemVisualizer.serviceContainer.Resolve(EventMediator);
+    }
+
     /* -------------------------------------------------------------------------- */
     /*                                Event Methods                               */
     /* -------------------------------------------------------------------------- */
@@ -23,6 +33,14 @@ class TimeControlRevealButton extends TimeControlButton {
         this.setState((prevState) => ({
             isHidden: !prevState.isHidden
         }));
+    }
+
+    HandleMouseEnter(event) {
+        GameManager.gameObserver.Dispatch("OnInterfaceEnter");
+    }
+
+    HandleMouseExit(event) {
+        GameManager.gameObserver.Dispatch("OnInterfaceExit");
     }
 
     // Tech-Debt: #152
@@ -46,20 +64,18 @@ class TimeControlRevealButton extends TimeControlButton {
     }
 
     /* -------------------------------------------------------------------------- */
-    /*                              Lifecycle Methods                             */
+    /*                                    View                                    */
     /* -------------------------------------------------------------------------- */
-
-    componentDidMount() {
-        GameManager.gameObserver.Subscribe("OnMouseDown", this.OnMouseDown.bind(this));
-        this.eventMediator = SolarSystemVisualizer.serviceContainer.Resolve(EventMediator);
-    }
 
     render() {
         const caretUpIconClassName = `time-control-icon ${this.state.isHidden ? "" : "hidden"}`;
         const caretDownIconClassName = `time-control-icon ${this.state.isHidden ? "hidden" : ""}`;
 
         return (
-            <div className="time-control-reveal-group">
+            <div
+                className="time-control-reveal-group"
+                onMouseEnter={this.HandleMouseEnter.bind(this)}
+                onMouseLeave={this.HandleMouseExit.bind(this)}>
                 <button className="time-control-expand-button" onClick={this.HandleClick.bind(this)}>
                     <FontAwesomeIcon icon={faCaretUp} className={caretUpIconClassName}/>
                     <p className="time-control-expand-button-text ">Time Controls</p>